Re-organize code of frame allocator's pools
